requirements.txt
Django
django-countries
django-debug-toolbar
django-import-export
django-jet
Pillow
dj-static